Fixed File Structure
Organised files into correct folders
