+++ src/components/flashcards/LibrarySideBar.jsx
@@ -1,12 +1,7 @@
 import React from "react";
-import { useState, useEffect } from "react";
-import { doc, collection, getDocs } from "firebase/firestore";
-import { db } from "../config/firebase";
-import useAuthStore from "../store/authStore";
-import useLibraryStore from "../store/libraryStore";
-import useGetDeckList from "../hooks/useGetDeckList";
-import "../styles/Library.css";
-import useGetSharedDeckList from "../hooks/useGetSharedDeckList";
+import useGetDeckList from "../../hooks/useGetDeckList";
+import "../../styles/Library.css";
+import useGetSharedDeckList from "../../hooks/useGetSharedDeckList";
 //Library is a collection containing the docs of decks
 const LibrarySideBar = () => {
   const deckList = useGetDeckList();
