+++ src/components/profile/Header.jsx
@@ -1,9 +1,8 @@
 import React from "react";
-import "./Header.css";
-import profile from "./assets/profile.png";
-import { auth, googleProvider } from "./config/firebase.js";
+import "../../styles/Header.css";
+import { auth, googleProvider } from "../../config/firebase.js";
 import { signOut } from "firebase/auth";
-import useAuthStore from "./store/authStore.js";
+import useAuthStore from "../../store/authStore.js";
 
 const handleLogout = async () => {
   await signOut(auth, googleProvider);
