+++ src/main.jsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import App from "./App.jsx";
-import Home from "./Home";
-import Library from "./Library";
-import DeckPage from "./DeckPage";
+import Home from "./pages/Home.jsx";
+import Library from "./pages/Library.jsx";
+import DeckPage from "./pages/DeckPage.jsx";
 import ProfilePage from "./pages/ProfilePage";
 import "./styles/index.css";
